[#143] feat: 현재 screenName 가져오는 ViewController 익스텐션 구현
